Oppdater README med modulstruktur
